Move packages PBF and PBEM to sub-packages of 'posted.game'
