+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/editors/EmailSenderEditor.java
@@ -4,7 +4,7 @@ import com.google.common.base.Ascii;
 import games.strategy.engine.ClientFileSystemHelper;
 import games.strategy.engine.data.properties.GameProperties;
 import games.strategy.engine.framework.GameRunner;
-import games.strategy.engine.pbem.IEmailSender;
+import games.strategy.engine.posted.game.pbem.IEmailSender;
 import games.strategy.triplea.settings.ClientSetting;
 import java.awt.GridBagConstraints;
 import java.awt.Insets;
