+++ game-core/src/main/java/games/strategy/engine/posted/game/pbem/PbemMessagePoster.java
@@ -1,11 +1,11 @@
-package games.strategy.engine.pbem;
+package games.strategy.engine.posted.game.pbem;
 
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.GamePlayer;
 import games.strategy.engine.data.properties.GameProperties;
 import games.strategy.engine.framework.GameDataFileUtils;
 import games.strategy.engine.history.IDelegateHistoryWriter;
-import games.strategy.engine.pbf.IForumPoster;
+import games.strategy.engine.posted.game.pbf.IForumPoster;
 import games.strategy.triplea.delegate.remote.IAbstractForumPosterDelegate;
 import games.strategy.triplea.ui.TripleAFrame;
 import games.strategy.triplea.ui.history.HistoryLog;
