+++ game-core/src/main/java/games/strategy/engine/posted/game/pbf/AxisAndAlliesForumPoster.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.pbf;
+package games.strategy.engine.posted.game.pbf;
 
 import games.strategy.triplea.UrlConstants;
