+++ game-core/src/main/java/games/strategy/triplea/delegate/remote/IAbstractForumPosterDelegate.java
@@ -2,7 +2,7 @@ package games.strategy.triplea.delegate.remote;
 
 import games.strategy.engine.delegate.IDelegate;
 import games.strategy.engine.message.IRemote;
-import games.strategy.engine.pbem.PbemMessagePoster;
+import games.strategy.engine.posted.game.pbem.PbemMessagePoster;
 
 /**
  * Logic for posting a save game to a forum. Supplements other game logic at points where it makes
